elearning-system/backend/requirements-prod.txt
# Include base requirements
-r requirements.txt

# Production-specific packages
gunicorn==21.2.0
gevent==23.7.0
django-storages==1.14.2
boto3==1.34.0
sentry-sdk==1.38.0

# Database URL parsing
dj-database-url==2.1.0

# Security and monitoring
django-ratelimit==4.1.0

# Performance
django-redis==5.4.0

# Health checks
django-health-check==3.17.0

# Production utilities
python-dotenv==1.0.0